+++ src/MapBBCodeUI.Editor.js
@@ -146,6 +146,8 @@ window.MapBBCode.include({
 
         var map = L.map(mapDiv, L.extend({}, { zoomControl: false }, this.options.leafletOptions));
         map.addControl(new L.Control.Zoom({ zoomInTitle: this.strings.zoomInTitle, zoomOutTitle: this.strings.zoomOutTitle }));
+        if( map.attributionControl )
+            map.attributionControl.setPrefix('<a href="http://mapbbcode.org" title="A library for [map] bbcode parsing, editing and displaying">MapBBCode</a>');
         if( L.Control.Search )
             map.addControl(new L.Control.Search({ title: this.strings.searchTitle }));
         this._addLayers(map);